+++ src/lib.rs
@@ -24,3 +24,5 @@
 pub mod attestation_client;
 #[cfg(feature = "protected-attestor")]
 pub mod attestor;
+#[cfg(target_os = "linux")]
+pub mod observation_collector;